create password reset view; extract eamil validation function to util class
