+++ iphone/UserLoginAndRegistration/UserLoginAndRegistration/RegisterPageViewController.swift
@@ -5,7 +5,7 @@
 //  Created by Chaoran Wang on 4/23/16.
 //  Copyright © 2016 food. All rights reserved.
 //
-
+import Foundation
 import UIKit
 
 class RegisterPageViewController: UIViewController {
@@ -46,7 +46,7 @@ class RegisterPageViewController: UIViewController {
             return;
         }
         
-        if (!isValidEmail(userEmail)){
+        if (!Utils.isValidEmail(userEmail)){
             displayMyAlertMessage("Please enter a valid email!");
             return;
         }
@@ -96,6 +96,9 @@ class RegisterPageViewController: UIViewController {
         task.resume();
     }
 
+    @IBAction func iHaveAnAccountButtonTapped(sender: UIButton) {
+        self.dismissViewControllerAnimated(true, completion: nil);
+    }
     
     func displayMyAlertMessage(userMessage:String) {
         let myAlert = UIAlertController(title: "Alert", message: userMessage, preferredStyle: UIAlertControllerStyle.Alert);
@@ -104,18 +107,6 @@ class RegisterPageViewController: UIViewController {
         self.presentViewController(myAlert, animated: true, completion: nil);
     }
     
-    func isValidEmail(testStr:String) -> Bool {
-        print("validate emilId: \(testStr)")
-        let emailRegEx = "^(?:(?:(?:(?: )*(?:(?:(?:\\t| )*\\r\\n)?(?:\\t| )+))+(?: )*)|(?: )+)?(?:(?:(?:[-A-Za-z0-9!#$%&’*+/=?^_'{|}~]+(?:\\.[-A-Za-z0-9!#$%&’*+/=?^_'{|}~]+)*)|(?:\"(?:(?:(?:(?: )*(?:(?:[!#-Z^-~]|\\[|\\])|(?:\\\\(?:\\t|[ -~]))))+(?: )*)|(?: )+)\"))(?:@)(?:(?:(?:[A-Za-z0-9](?:[-A-Za-z0-9]{0,61}[A-Za-z0-9])?)(?:\\.[A-Za-z0-9](?:[-A-Za-z0-9]{0,61}[A-Za-z0-9])?)*)|(?:\\[(?:(?:(?:(?:(?:[0-9]|(?:[1-9][0-9])|(?:1[0-9][0-9])|(?:2[0-4][0-9])|(?:25[0-5]))\\.){3}(?:[0-9]|(?:[1-9][0-9])|(?:1[0-9][0-9])|(?:2[0-4][0-9])|(?:25[0-5]))))|(?:(?:(?: )*[!-Z^-~])*(?: )*)|(?:[Vv][0-9A-Fa-f]+\\.[-A-Za-z0-9._~!$&'()*+,;=:]+))\\])))(?:(?:(?:(?: )*(?:(?:(?:\\t| )*\\r\\n)?(?:\\t| )+))+(?: )*)|(?: )+)?$"
-        let emailTest = NSPredicate(format:"SELF MATCHES %@", emailRegEx)
-        let result = emailTest.evaluateWithObject(testStr)
-        return result
-    }
-    
-
-    @IBAction func iHaveAnAccountButtonTapped(sender: UIButton) {
-        self.dismissViewControllerAnimated(true, completion: nil);
-    }
     /*
     // MARK: - Navigation
