+++ iphone/UserLoginAndRegistration/UserLoginAndRegistration/ResetPasswordViewController.swift
@@ -1,0 +1,43 @@
+//
+//  ResetPasswordViewController.swift
+//  UserLoginAndRegistration
+//
+//  Created by Chaoran Wang on 4/27/16.
+//  Copyright © 2016 food. All rights reserved.
+//
+
+import UIKit
+
+class ResetPasswordViewController: UIViewController {
+
+    @IBOutlet weak var userEmailTextField: UITextField!
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view.
+    }
+
+    override func didReceiveMemoryWarning() {
+        super.didReceiveMemoryWarning()
+        // Dispose of any resources that can be recreated.
+    }
+    
+    
+    @IBAction func confirmResetButtonTapped(sender: UIButton) {
+        let userEmail = userEmailTextField.text!;
+        if(userEmail.isEmpty) { return; }
+        
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepareForSegue(segue: UIStoryboardSegue, sender: AnyObject?) {
+        // Get the new view controller using segue.destinationViewController.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
